requirements.txt
sqlalchemy
psycopg2-binary
redis
flask
async-timeout
load_dotenv